[FIX] website: restore default h1 color in customize dialog
Since the bootstrap update that was merged with
https://github.com/odoo/odoo/commit/c116690b2da4ce3def8f6d7d710f8706f2bbd232
the customize theme dialog was showing transparent colors for the
headings. This was because bootstrap changed the value of the related
variable from 'inherit' to 'null' to avoid an useless css rule if the
headings are supposed to have the same color as the body text.

closes odoo/odoo#31909

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>