[FIX] hr_holidays: use api model instead of multi for dashboard
Before aff0b62e8, the get_days_all_request method was called with the
context as an argument. As a consequence, because of the api.multi, the
first arg is used as the ids of the record. As the id is not used in
this method, it was not harmful. With the above mentioned fix, there is
no more argument. This causes a traceback when clicking on menu "My Time
Off/Dashboard".

With this commit, the api.model is used instead, that way, the first arg
is not considered as ids.

closes odoo/odoo#32119

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>