[FIX] base: Add missing parent_id in res.partner form
Steps to reproduce:
- Install hr and base_address_city
- Edit an employee
- In private info, set a private address
- Open private address form from the employee form

Traceback because parent_id is undefined.

The base_address_city module override a part of res.partner
view form to change the city readonly modifier.
This override adds a domain in which appears parent_id,
but the field is not present in the view.

Related commit a7274b3c1cc4dc5b539ad1f071ea309020c26be6

closes odoo/odoo#32172

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>