[IMP] website_slides: Fix slides displayed by category
This commit fixes an issue where only three slides would be
displayed instead of four in a category of the course page.

This was due to the promoted slide being ignored in its own
category.

After this commit, the promoted slide will appear in the promoted
slot AND in its category

TaskID: 1952064
PR: #32229